Rename unk_0206C784 and overlay93 to boat_cutscene
