+++ include/boat_cutscene/boat_cutscene.h
@@ -1,5 +1,5 @@
-#ifndef POKEPLATINUM_UNK_0206C784_H
-#define POKEPLATINUM_UNK_0206C784_H
+#ifndef POKEPLATINUM_BOAT_CUTSCENE_H
+#define POKEPLATINUM_BOAT_CUTSCENE_H
 
 #include "field/field_system_decl.h"
 #include "overlay005/model_attributes.h"
@@ -26,4 +26,4 @@ typedef struct BoatTravelCutscene {
 
 void FieldSystem_PlayBoatCutscene(FieldSystem *fieldSystem, const u8 startDir, const u8 endDir, const int mapID, const int x, const int z);
 
-#endif // POKEPLATINUM_UNK_0206C784_H
+#endif // POKEPLATINUM_BOAT_CUTSCENE_H
