+++ include/boat_cutscene/ship_demo.h
@@ -1,0 +1,10 @@
+#ifndef POKEPLATINUM_BOAT_CUTSCENE_SHIP_DEMO_H
+#define POKEPLATINUM_BOAT_CUTSCENE_SHIP_DEMO_H
+
+#include "overlay_manager.h"
+
+int BoatTravelCutscene_ShipDemo_Init(ApplicationManager *appMan, int *state);
+int BoatTravelCutscene_ShipDemo_Main(ApplicationManager *appMan, int *state);
+int BoatTravelCutscene_ShipDemo_Exit(ApplicationManager *appMan, int *state);
+
+#endif // POKEPLATINUM_BOAT_CUTSCENE_SHIP_DEMO_H
