+++ include/boat_cutscene/ship_demo_pl.h
@@ -1,0 +1,10 @@
+#ifndef POKEPLATINUM_BOAT_CUTSCENE_SHIP_DEMO_PL_H
+#define POKEPLATINUM_BOAT_CUTSCENE_SHIP_DEMO_PL_H
+
+#include "overlay_manager.h"
+
+int BoatTravelCutscene_ShipDemoPl_Init(ApplicationManager *appMan, int *state);
+int BoatTravelCutscene_ShipDemoPl_Main(ApplicationManager *appMan, int *state);
+int BoatTravelCutscene_ShipDemoPl_Exit(ApplicationManager *appMan, int *state);
+
+#endif // POKEPLATINUM_BOAT_CUTSCENE_SHIP_DEMO_PL_H
