--- include/overlay093/ov93_021D0D80.h
@@ -1,10 +1,0 @@
-#ifndef POKEPLATINUM_OV93_021D0D80_H
-#define POKEPLATINUM_OV93_021D0D80_H
-
-#include "overlay_manager.h"
-
-int BoatTravelCutscene_ShipDemo_Init(ApplicationManager *appMan, int *unused);
-int BoatTravelCutscene_ShipDemo_Main(ApplicationManager *appMan, int *param1);
-int BoatTravelCutscene_ShipDemo_Exit(ApplicationManager *appMan, int *param1);
-
-#endif // POKEPLATINUM_OV93_021D0D80_H
