--- include/overlay093/ov93_021D111C.h
@@ -1,10 +1,0 @@
-#ifndef POKEPLATINUM_OV93_021D111C_H
-#define POKEPLATINUM_OV93_021D111C_H
-
-#include "overlay_manager.h"
-
-int BoatTravelCutscene_ShipDemoPl_Init(ApplicationManager *appMan, int *param1);
-int BoatTravelCutscene_ShipDemoPl_Main(ApplicationManager *appMan, int *param1);
-int BoatTravelCutscene_ShipDemoPl_Exit(ApplicationManager *appMan, int *param1);
-
-#endif // POKEPLATINUM_OV93_021D111C_H
