+++ platinum.us/sbins_shared.sha1
@@ -92,7 +92,7 @@ de8a847bff8c343d69b853a215e6ee775ef2ef96 *overlay89.sbin
 84ce4bc9949f697c12ae6598ef74a7e3f60b1531 *overlay90.sbin
 2c26cf42ca34629dfc7cc491820e2a309e57eeeb *overlay91.sbin
 87c3ffad8da5d31174a542832d80bf1855c3225c *overlay92.sbin
-dabd69f841fce75c91c4236925757150990af289 *overlay93.sbin
+dabd69f841fce75c91c4236925757150990af289 *boat_cutscene.sbin
 3410ea804433b4a797932d1999fbb2429611ae56 *overlay95.sbin
 af7de8e345ff25de55130069f313b57e85d72a4f *overlay96.sbin
 138a69ea796d35f244daa36a7e9962ea2d0322e0 *overlay97.sbin
