+++ src/boat_cutscene/ship_demo_pl.c
@@ -1,10 +1,11 @@
-#include "overlay093/ov93_021D0D80.h"
+#include "boat_cutscene/ship_demo_pl.h"
 
 #include <nitro.h>
 #include <string.h>
 
 #include "struct_defs/struct_02099F80.h"
 
+#include "boat_cutscene/boat_cutscene.h"
 #include "overlay009/camera_configuration.h"
 
 #include "camera.h"
@@ -17,32 +18,31 @@
 #include "sound_playback.h"
 #include "system.h"
 #include "unk_0202419C.h"
-#include "unk_0206C784.h"
 #include "unk_0208C098.h"
 
-#define BOAT_TRAVEL_CUTSCENE_SHIP_DEMO_NUM_ANIMATIONS 4
+#define BOAT_TRAVEL_CUTSCENE_SHIP_DEMO_PL_NUM_ANIMATIONS 3
 
-typedef struct BoatTravelCutscene_ShipDemo {
+typedef struct BoatTravelCutscene_ShipDemoPl {
     Camera *camera;
     NNSG3dRenderObj renderObj;
     NNSG3dResMdl *model;
     NNSG3dResFileHeader *modelFile;
-    void *animationFiles[BOAT_TRAVEL_CUTSCENE_SHIP_DEMO_NUM_ANIMATIONS];
+    void *animationFiles[BOAT_TRAVEL_CUTSCENE_SHIP_DEMO_PL_NUM_ANIMATIONS];
     NNSFndAllocator allocator;
-    NNSG3dAnmObj *anmObjs[BOAT_TRAVEL_CUTSCENE_SHIP_DEMO_NUM_ANIMATIONS];
+    NNSG3dAnmObj *anmObjs[BOAT_TRAVEL_CUTSCENE_SHIP_DEMO_PL_NUM_ANIMATIONS];
     VecFx32 unused;
     u8 startDir;
     u8 firstAnimationFinished;
-} BoatTravelCutscene_ShipDemo;
+} BoatTravelCutscene_ShipDemoPl;
 
-typedef struct BoatTravelCutscene_ShipDemo_NarcMemberIndexes {
+typedef struct BoatTravelCutscene_ShipDemoPl_NarcMemberIndexes {
     u16 modelMemberIndex;
-    u16 animationMemberIndexes[BOAT_TRAVEL_CUTSCENE_SHIP_DEMO_NUM_ANIMATIONS];
-} BoatTravelCutscene_ShipDemo_NarcMemberIndexes;
+    u16 animationMemberIndexes[BOAT_TRAVEL_CUTSCENE_SHIP_DEMO_PL_NUM_ANIMATIONS];
+} BoatTravelCutscene_ShipDemoPl_NarcMemberIndexes;
 
-static void BoatTravelCutscene_ShipDemo_Init3D(void);
-static void BoatTravelCutscene_ShipDemo_SetGXBanks(void);
-static void BoatTravelCutscene_ShipDemo_LoadAnimations(BoatTravelCutscene_ShipDemo *cutsceneData);
+static void BoatTravelCutscene_ShipDemoPl_Init3D(void);
+static void BoatTravelCutscene_ShipDemoPl_SetGXBanks(void);
+static void BoatTravelCutscene_ShipDemoPl_LoadAnimations(BoatTravelCutscene_ShipDemoPl *cutsceneData);
 
 static const CameraConfiguration cameraConfig = {
     .distance = BOAT_TRAVEL_CUTSCENE_CAMERA_BASE_DISTANCE,
@@ -65,35 +65,35 @@ static const GXRgb edgeColorTable[8] = {
     GX_RGB(4, 4, 4)
 };
 
-static BoatTravelCutscene_ShipDemo_NarcMemberIndexes narcMemberIndexes[4] = {
+static BoatTravelCutscene_ShipDemoPl_NarcMemberIndexes narcMemberIndexes[4] = {
     {
-        2,
-        { 0, 1, 3, 4 },
+        1,
+        { 3, 2, 0 },
     },
     {
-        7,
-        { 5, 6, 8, 9 },
+        5,
+        { 7, 6, 4 },
     },
     {
-        12,
-        { 10, 11, 13, 14 },
+        1,
+        { 3, 2, 0 },
     },
     {
-        17,
-        { 15, 16, 18, 19 },
+        1,
+        { 3, 2, 0 },
     },
 };
 
-int BoatTravelCutscene_ShipDemo_Init(ApplicationManager *appMan, int *state)
+int BoatTravelCutscene_ShipDemoPl_Init(ApplicationManager *appMan, int *state)
 {
     u8 lightID;
-    BoatTravelCutscene_ShipDemo *cutsceneData;
+    BoatTravelCutscene_ShipDemoPl *cutsceneData;
     BoatTravelCutscene *taskEnv;
 
     Heap_Create(HEAP_ID_APPLICATION, HEAP_ID_72, BOAT_TRAVEL_CUTSCENE_HEAP_SIZE);
 
-    cutsceneData = ApplicationManager_NewData(appMan, sizeof(BoatTravelCutscene_ShipDemo), HEAP_ID_72);
-    memset(cutsceneData, 0, sizeof(BoatTravelCutscene_ShipDemo));
+    cutsceneData = ApplicationManager_NewData(appMan, sizeof(BoatTravelCutscene_ShipDemoPl), HEAP_ID_72);
+    memset(cutsceneData, 0, sizeof(BoatTravelCutscene_ShipDemoPl));
 
     taskEnv = ApplicationManager_Args(appMan);
 
@@ -101,8 +101,8 @@ int BoatTravelCutscene_ShipDemo_Init(ApplicationManager *appMan, int *state)
     cutsceneData->firstAnimationFinished = FALSE;
     cutsceneData->camera = Camera_Alloc(HEAP_ID_72);
 
-    BoatTravelCutscene_ShipDemo_Init3D();
-    BoatTravelCutscene_ShipDemo_LoadAnimations(cutsceneData);
+    BoatTravelCutscene_ShipDemoPl_Init3D();
+    BoatTravelCutscene_ShipDemoPl_LoadAnimations(cutsceneData);
 
     VecFx32 cameraTarget = { 0, 0, 0 };
 
@@ -122,7 +122,7 @@ int BoatTravelCutscene_ShipDemo_Init(ApplicationManager *appMan, int *state)
     return TRUE;
 }
 
-int BoatTravelCutscene_ShipDemo_Main(ApplicationManager *appMan, int *state)
+int BoatTravelCutscene_ShipDemoPl_Main(ApplicationManager *appMan, int *state)
 {
     u8 animIndex;
     BOOL unused;
@@ -147,7 +147,7 @@ int BoatTravelCutscene_ShipDemo_Main(ApplicationManager *appMan, int *state)
         0,
         0
     };
-    BoatTravelCutscene_ShipDemo *cutsceneData = ApplicationManager_Data(appMan);
+    BoatTravelCutscene_ShipDemoPl *cutsceneData = ApplicationManager_Data(appMan);
 
     switch (*state) {
     case BOAT_TRAVEL_CUTSCENE_STATE_SFX:
@@ -167,7 +167,7 @@ int BoatTravelCutscene_ShipDemo_Main(ApplicationManager *appMan, int *state)
         }
     }
 
-    for (animIndex = 0; animIndex < BOAT_TRAVEL_CUTSCENE_SHIP_DEMO_NUM_ANIMATIONS; animIndex++) {
+    for (animIndex = 0; animIndex < BOAT_TRAVEL_CUTSCENE_SHIP_DEMO_PL_NUM_ANIMATIONS; animIndex++) {
         if (cutsceneData->anmObjs[animIndex]->frame + FX32_ONE < NNS_G3dAnmObjGetNumFrame(cutsceneData->anmObjs[animIndex])) {
             cutsceneData->anmObjs[animIndex]->frame += FX32_ONE;
         }
@@ -178,15 +178,15 @@ int BoatTravelCutscene_ShipDemo_Main(ApplicationManager *appMan, int *state)
     Easy3D_DrawRenderObj(&cutsceneData->renderObj, &position, &rotation, &scale);
     G3_RequestSwapBuffers(GX_SORTMODE_MANUAL, GX_BUFFERMODE_W);
 
-    return 0;
+    return FALSE;
 }
 
-int BoatTravelCutscene_ShipDemo_Exit(ApplicationManager *appMan, int *state)
+int BoatTravelCutscene_ShipDemoPl_Exit(ApplicationManager *appMan, int *state)
 {
     u8 animIndex;
-    BoatTravelCutscene_ShipDemo *cutsceneData = ApplicationManager_Data(appMan);
+    BoatTravelCutscene_ShipDemoPl *cutsceneData = ApplicationManager_Data(appMan);
 
-    for (animIndex = 0; animIndex < BOAT_TRAVEL_CUTSCENE_SHIP_DEMO_NUM_ANIMATIONS; animIndex++) {
+    for (animIndex = 0; animIndex < BOAT_TRAVEL_CUTSCENE_SHIP_DEMO_PL_NUM_ANIMATIONS; animIndex++) {
         NNS_G3dFreeAnmObj(&cutsceneData->allocator, cutsceneData->anmObjs[animIndex]);
         Heap_Free(cutsceneData->animationFiles[animIndex]);
     }
@@ -200,7 +200,7 @@ int BoatTravelCutscene_ShipDemo_Exit(ApplicationManager *appMan, int *state)
     return TRUE;
 }
 
-static void BoatTravelCutscene_ShipDemo_Init3D(void)
+static void BoatTravelCutscene_ShipDemoPl_Init3D(void)
 {
     SetVBlankCallback(NULL, NULL);
     DisableHBlank();
@@ -210,7 +210,7 @@ static void BoatTravelCutscene_ShipDemo_Init3D(void)
     GX_SetVisiblePlane(0);
     GXS_SetVisiblePlane(0);
 
-    BoatTravelCutscene_ShipDemo_SetGXBanks();
+    BoatTravelCutscene_ShipDemoPl_SetGXBanks();
     Easy3D_Init(HEAP_ID_72);
 
     G3X_EdgeMarking(TRUE);
@@ -219,7 +219,7 @@ static void BoatTravelCutscene_ShipDemo_Init3D(void)
     GXLayers_SwapDisplay();
 }
 
-static void BoatTravelCutscene_ShipDemo_SetGXBanks(void)
+static void BoatTravelCutscene_ShipDemoPl_SetGXBanks(void)
 {
     UnkStruct_02099F80 banks = {
         GX_VRAM_BG_128_C,
@@ -237,12 +237,12 @@ static void BoatTravelCutscene_ShipDemo_SetGXBanks(void)
     GXLayers_SetBanks(&banks);
 }
 
-static void BoatTravelCutscene_ShipDemo_LoadAnimations(BoatTravelCutscene_ShipDemo *cutsceneData)
+static void BoatTravelCutscene_ShipDemoPl_LoadAnimations(BoatTravelCutscene_ShipDemoPl *cutsceneData)
 {
     u8 animIndex;
     NNSG3dResTex *texture;
     void *animation;
-    NARC *narc = NARC_ctor(NARC_INDEX_ARC__SHIP_DEMO, HEAP_ID_72);
+    NARC *narc = NARC_ctor(NARC_INDEX_ARC__SHIP_DEMO_PL, HEAP_ID_72);
     Heap_FndInitAllocatorForExpHeap(&cutsceneData->allocator, HEAP_ID_72, 4);
 
     cutsceneData->modelFile = NARC_AllocAndReadWholeMember(narc, narcMemberIndexes[cutsceneData->startDir].modelMemberIndex, HEAP_ID_72);
@@ -254,7 +254,7 @@ static void BoatTravelCutscene_ShipDemo_LoadAnimations(BoatTravelCutscene_ShipDe
     NNS_G3dMdlUseGlbSpec(cutsceneData->model);
     NNS_G3dMdlUseGlbEmi(cutsceneData->model);
 
-    for (animIndex = 0; animIndex < BOAT_TRAVEL_CUTSCENE_SHIP_DEMO_NUM_ANIMATIONS; animIndex++) {
+    for (animIndex = 0; animIndex < BOAT_TRAVEL_CUTSCENE_SHIP_DEMO_PL_NUM_ANIMATIONS; animIndex++) {
         cutsceneData->animationFiles[animIndex] = NARC_AllocAndReadWholeMember(narc, narcMemberIndexes[cutsceneData->startDir].animationMemberIndexes[animIndex], HEAP_ID_72);
         animation = NNS_G3dGetAnmByIdx(cutsceneData->animationFiles[animIndex], 0);
         cutsceneData->anmObjs[animIndex] = NNS_G3dAllocAnmObj(&cutsceneData->allocator, animation, cutsceneData->model);
