+++ src/meson.build
@@ -258,7 +258,7 @@ pokeplatinum_c = files(
     'trainer_card_badge_dirt.c',
     'roaming_pokemon.c',
     'unk_0206C660.c',
-    'unk_0206C784.c',
+    'boat_cutscene/boat_cutscene.c',
     'unk_0206CCB0.c',
     'unk_0206F314.c',
     'unk_02070428.c',
@@ -848,8 +848,8 @@ pokeplatinum_c = files(
     'overlay090/ov90_021D0D80.c',
     'overlay091/ov91_021D0D80.c',
     'overlay092/ov92_021D0D80.c',
-    'overlay093/ov93_021D0D80.c',
-    'overlay093/ov93_021D111C.c',
+    'boat_cutscene/ship_demo.c',
+    'boat_cutscene/ship_demo_pl.c',
     'overlay094/ov94_0223B140.c',
     'overlay094/ov94_0223BCB0.c',
     'overlay094/ov94_0223C610.c',
