+++ src/unk_0203D1B8.c
@@ -39,6 +39,8 @@
 #include "applications/pokedex/pokedex_main.h"
 #include "applications/pokemon_summary_screen/main.h"
 #include "battle/ov16_0223B140.h"
+#include "boat_cutscene/ship_demo.h"
+#include "boat_cutscene/ship_demo_pl.h"
 #include "choose_starter/choose_starter_app.h"
 #include "dw_warp/dw_warp.h"
 #include "field/field_system.h"
@@ -67,8 +69,6 @@
 #include "overlay090/struct_ov90_021D0D80.h"
 #include "overlay091/ov91_021D0D80.h"
 #include "overlay092/ov92_021D0D80.h"
-#include "overlay093/ov93_021D0D80.h"
-#include "overlay093/ov93_021D111C.h"
 #include "overlay094/ov94_0223BCB0.h"
 #include "overlay095/ov95_02246C20.h"
 #include "overlay096/ov96_0223B6A0.h"
@@ -165,7 +165,7 @@ FS_EXTERN_OVERLAY(overlay88);
 FS_EXTERN_OVERLAY(overlay90);
 FS_EXTERN_OVERLAY(overlay91);
 FS_EXTERN_OVERLAY(overlay92);
-FS_EXTERN_OVERLAY(overlay93);
+FS_EXTERN_OVERLAY(boat_cutscene);
 FS_EXTERN_OVERLAY(overlay94);
 FS_EXTERN_OVERLAY(overlay95);
 FS_EXTERN_OVERLAY(overlay96);
@@ -1475,13 +1475,13 @@ void sub_0203E284(FieldSystem *fieldSystem, UnkStruct_020997B8 *param1)
 
 void FieldTask_PlayBoatTravelCutscene_ShipDemo(FieldSystem *fieldSystem, void *taskEnv)
 {
-    FS_EXTERN_OVERLAY(overlay93);
+    FS_EXTERN_OVERLAY(boat_cutscene);
 
     const ApplicationManagerTemplate appTemplate = {
         .init = BoatTravelCutscene_ShipDemo_Init,
         .main = BoatTravelCutscene_ShipDemo_Main,
         .exit = BoatTravelCutscene_ShipDemo_Exit,
-        .overlayID = FS_OVERLAY_ID(overlay93)
+        .overlayID = FS_OVERLAY_ID(boat_cutscene)
     };
 
     FieldSystem_StartChildProcess(fieldSystem, &appTemplate, taskEnv);
@@ -1489,13 +1489,13 @@ void FieldTask_PlayBoatTravelCutscene_ShipDemo(FieldSystem *fieldSystem, void *t
 
 void FieldTask_PlayBoatTravelCutscene_ShipDemoPl(FieldSystem *fieldSystem, void *taskEnv)
 {
-    FS_EXTERN_OVERLAY(overlay93);
+    FS_EXTERN_OVERLAY(boat_cutscene);
 
     const ApplicationManagerTemplate appTemplate = {
         .init = BoatTravelCutscene_ShipDemoPl_Init,
         .main = BoatTravelCutscene_ShipDemoPl_Main,
         .exit = BoatTravelCutscene_ShipDemoPl_Exit,
-        .overlayID = FS_OVERLAY_ID(overlay93)
+        .overlayID = FS_OVERLAY_ID(boat_cutscene)
     };
 
     FieldSystem_StartChildProcess(fieldSystem, &appTemplate, taskEnv);
